CliffordNumbers

CliffordNumbers.jl is a package that provides fully static multivectors (Clifford numbers) in arbitrary dimensions and metrics. While in many cases, sparse representations of multivectors are more efficient, for spaces of low dimension, dense static representations may provide a performance and convenience advantage.

CliffordNumbers.APS

CliffordNumbers.AbstractCliffordNumber

CliffordNumbers.BaseNumber

CliffordNumbers.BitIndex

CliffordNumbers.BitIndex

CliffordNumbers.BitIndices

CliffordNumbers.BitIndices

CliffordNumbers.CliffordNumber

CliffordNumbers.KVector

CliffordNumbers.QuadraticForm

CliffordNumbers.STA

Base.:*

Base.:~

Base.abs

Base.abs2

Base.conj

Base.exp

Base.real

Base.reverse

Base.sign

CliffordNumbers.CGA

CliffordNumbers.PGA

CliffordNumbers.VGA

CliffordNumbers._sort_with_parity!

CliffordNumbers.dot

CliffordNumbers.dual

CliffordNumbers.elementwise_product

CliffordNumbers.evil_number

CliffordNumbers.exppi

CliffordNumbers.exptau

CliffordNumbers.grade_involution

CliffordNumbers.hamming_number

CliffordNumbers.hestenes_product

CliffordNumbers.isevil

CliffordNumbers.isodious

CliffordNumbers.ispseudoscalar

CliffordNumbers.isscalar

CliffordNumbers.left_contraction

CliffordNumbers.normalize

CliffordNumbers.odious_number

CliffordNumbers.right_contraction

CliffordNumbers.sandwich

CliffordNumbers.scalar_product

CliffordNumbers.select_grade

CliffordNumbers.sign_of_mult

CliffordNumbers.signbit_of_mult

CliffordNumbers.signmask

CliffordNumbers.subscript_string

CliffordNumbers.undual

CliffordNumbers.versor_inverse

CliffordNumbers.wedge

CliffordNumbers.BaseNumber — Type

CliffordNumbers.BaseNumber

Union of subtypes of Number provided in the Julia Base module: Real and Complex. This encompasses all types that may be used to construct a CliffordNumber.

CliffordNumbers.APS — Type

APS

The algebra of physical space, Cl(3,0,0). An alias for QuadraticForm{3,0,0}.

CliffordNumbers.AbstractCliffordNumber — Type

AbstractCliffordNumber{Q,T}

An element of a Clifford algebra, often referred to as a multivector, with quadratic form Q and element type T.

CliffordNumbers.BitIndex — Type

BitIndex{Q<:QuadraticForm}

A representation of an index corresponding to a basis blade of the geometric algebra with quadratic form Q.

CliffordNumbers.BitIndex — Method

BitIndex{Q}(x::Integer...)
BitIndex{Q}(v::AbstractVector{<:Integer})

Constructs a BitIndex{Q} from a list of integers that represent the basis vectors of the space.

This package uses a lexicographic convention for basis blades: in the algebra of physical space, the basis bivectors are {e₁e₂, e₁e₃, e₂e₃}. The sign of the BitIndex{Q} is negative when the parity of the basis vector permutation is odd.

CliffordNumbers.BitIndices — Type

BitIndices{Q<:QuadraticForm}

Represents a range of valid BitIndex objects for a given quadratic form.

For sparse representations, such as KVector{K...}, this is not the most efficient way to iterate through all elements, as it includes indices that are known to be zero.

CliffordNumbers.BitIndices — Method

BitIndices(x::AbstractCliffordNumber{Q}) -> BitIndices{Q}()
BitIndices(T::Type{<:AbstractCliffordNumber{Q}}) -> BitIndices{Q}()

Constructs a BitIndices object associated with a Clifford number or its type.

CliffordNumbers.CliffordNumber — Type

CliffordNumber{Q,T,L}

A dense multivector (or Clifford number), with quadratic form Q, element type T, and length L (which depends entirely on Q).

The coefficients are ordered by taking advantage of the natural binary structure of the basis. The grade of an element is given by the Hamming weight of its index. For the algebra of physical space, the order is: 1, e₁, e₂, e₁₂, e₃, e₁₃, e₂₃, e₁₂₃ = i. This order allows for more aggressive SIMD optimization when calculating the geometric product.

CliffordNumbers.KVector — Type

KVector{K,Q,T,L}

A multivector consisting only linear combinations of basis blades of grade K - in other words, a k-vector.

k-vectors have binomial(dimension(Q), K) components.

CliffordNumbers.QuadraticForm — Type

CliffordNumbers.QuadratricForm

Represents a quadratric with P dimensions which square to +1, Q dimensions which square to -1, and R dimensions which square to 0, in that order.

By convention, this type is used as a tag, and is never instantiated.

CliffordNumbers.STA — Type

STA

Spacetime algebra with a mostly negative signature (particle physicist's convention), Cl(1,3,0). An alias for QuadraticForm{1,3,0}.

The negative signature is used by default to distinguish this algebra from conformal geometric algebras, which use a mostly positive signature by convention.

Base.:* — Method

*(x::CliffordNumber{Q}, y::CliffordNumber{Q}) -> CliffordNumber{Q}

Calculates the geometric product between multivectors/Clifford numbers x and y which share the quadratic form Q.

Base.:~ — Method

reverse(x::CliffordNumber{Q,T}) -> CliffordNumber{Q,T}
~(x::CliffordNumber{Q,T}) -> CliffordNumber{Q,T}

Calculate the reverse of a Clifford number. This effectively reverses the products that form the basis blades, or in other words, reverses the order of the geometric product that resulted in x.

Base.abs — Method

abs2(x::CliffordNumber{Q}) -> CliffordNumber{Q}

Calculates the norm of x, equal to sqrt(scalar_product(x, ~x)).

Base.abs2 — Method

abs2(x::CliffordNumber{Q}) -> CliffordNumber{Q}

Calculates the squared norm of x, equal to scalar_product(x, ~x).

Base.conj — Method

conj(x::CliffordNumber{Q,T}) -> CliffordNumber{Q,T}

Calculates the Clifford conjugate of a Clifford number x. This is equal to grade_involution(reverse(x)).

Base.exp — Method

exp(x::CliffordNumber{Q}) -> CliffordNumber{Q,<:AbstractFloat}

Returns the natural exponential of a Clifford number.

For special cases where m squares to a scalar, the following shortcuts can be used to calculate exp(x):

When x^2 < 0: exp(x) === cos(abs(x)) + x * sin(abs(x)) / abs(x)

When x^2 > 0: exp(x) === cosh(abs(x)) + x * sinh(abs(x)) / abs(x)

When x^2 === 0: exp(x) == 1 + x

See also: exppi, exptau.

Base.real — Method

real(x::CliffordNumber{Q,T<:Real}) = T

Return the real (scalar) portion of a real Clifford number.

Base.reverse — Method

reverse(x::CliffordNumber{Q,T}) -> CliffordNumber{Q,T}
~(x::CliffordNumber{Q,T}) -> CliffordNumber{Q,T}

Calculate the reverse of Clifford number x. This effectively reverses the products that form the basis blades, or in other words, reverses the order of the geometric product that resulted in x.

Base.sign — Method

sign(::Type{QuadraticForm{P,Q,R}}, i::Integer) -> Int8

Gets the sign associated with dimension i of a quadratric form.

CliffordNumbers.CGA — Method

CGA(D) -> Type{QuadraticForm{D+1,1,0}}

Creates the type of a quadratic form associated with a conformal geometric algebra (CGA) of dimension D.

For reasons of type stability, avoid calling this function without constant arguments.

CliffordNumbers.PGA — Method

PGA(D) -> Type{QuadraticForm{D,0,1}}

Creates the type of a quadratic form associated with a projective geometric algebra (PGA) of dimension D.

For reasons of type stability, avoid calling this function without constant arguments.

CliffordNumbers.VGA — Method

VGA(D) -> Type{QuadraticForm{D,0,0}}

Creates the type of a quadratic form associated with a vector/vanilla geometric algebra (VGA) of dimension D.

For reasons of type stability, avoid calling this function without constant arguments.

CliffordNumbers._sort_with_parity! — Method

CliffordNumbers._sort_with_parity!(v::AbstractVector{<:Real}) -> Tuple{typeof(v),Bool}

Performs a parity-tracking insertion sort of v, which modifies v in place. The function returns a tuple containing v and the parity, which is true for an odd permutation and false for an even permutation. This is implemented with a modified insertion sort algorithm.

CliffordNumbers.dot — Method

dot(x::CliffordNumber{Q}, y::CliffordNumber{Q}) -> CliffordNumber{Q}

Calculates the dot product of x and y.

For basis blades A of grade m and B of grade n, the dot product is equal to the left contraction when m >= n and is equal to the right contraction when n >= m.

CliffordNumbers.dual — Method

dual(x::CliffordNumber) -> CliffordNumber

Calculates the dual of x, which is equal to the left contraction of x with the inverse of the pseudoscalar. However,

Note that the dual has some properties that depend on the dimension and quadratic form:

The inverse of the unit pseudoscalar depends on the dimension of the space. Therefore, the

periodicity of

If the metric is degenerate, the dual is not unique.

CliffordNumbers.elementwise_product — Method

CliffordAlgebra.elementwise_product(
    x::CliffordNumber{Q},
    y::CliffordNumber{Q},
    a::BitIndex{Q},
    b::BitIndex{Q}
)

Calculates the geometric product between the element of x indexed by a and the element of y indexed by b.

CliffordNumbers.evil_number — Method

CliffordNumbers.evil_number(n::Integer)

Returns the nth evil number, with the first evil number (n == 1) defined to be 0.

Evil numbers are numbers which have an even Hamming weight (sum of its binary digits).

CliffordNumbers.exppi — Method

exppi(x::CliffordNumber)

Returns the natural exponential of π * x with greater accuracy than exp(π * x) in the case where x^2 is a negative scalar.

See also: exp, exptau.

CliffordNumbers.exptau — Method

exptau(x::CliffordNumber)

Returns the natural exponential of 2π * x with greater accuracy than exp(2π * x) in the case where x^2 is a negative scalar.

See also: exp, exppi.

CliffordNumbers.grade_involution — Method

grade_involution(x::CliffordNumber{Q,T}) -> CliffordNumber{Q,T}

Calculates the grade involution of Clifford number x. This effectively multiplies all of the basis vectors of the space by -1, which makes elements of odd grade flip sign.

CliffordNumbers.hamming_number — Method

CliffordNumbers.hamming_number(w::Integer, n::Integer)

Gets the nth number with Hamming weight w. The first number with this Hamming weight (n = 1) is 2^w - 1.

Example

julia> CliffordNumbers.hamming_number(3, 2)
11

CliffordNumbers.hestenes_product — Method

hestenes_product(x::CliffordNumber{Q}, y::CliffordNumber{Q}) -> CliffordNumber{Q}

Returns the Hestenes product: this is equal to the dot product given by dot(x, y) but is equal to to zero when either x or y is a scalar.

CliffordNumbers.isevil — Method

CliffordNumbers.isevil(i::Integer) -> Bool

Determines whether a number is evil, meaning that its Hamming weight (sum of its binary digits) is even.

CliffordNumbers.isodious — Method

CliffordNumbers.isodious(i::Integer) -> Bool

Determines whether a number is odious, meaning that its Hamming weight (sum of its binary digits) is odd.

CliffordNumbers.ispseudoscalar — Method

ispseudoscalar(m::CliffordNumber)

Determines whether the Clifford number m is a pseudoscalar, meaning that it has no components with grades below the one equal to the dimension of the space.

CliffordNumbers.isscalar — Method

isscalar(m::CliffordNumber)

Determines whether the Clifford number m is a scalar, meaning that it has no components with grades above zero.

CliffordNumbers.left_contraction — Method

left_contraction(x::CliffordNumber{Q}, y::CliffordNumber{Q}) -> CliffordNumber{Q}

Calculates the left contraction of x and y.

For basis blades A of grade m and B of grade n, the left contraction is zero if n < m, otherwise it is grade_select(A*B, n-m).

CliffordNumbers.normalize — Method

normalize(x::CliffordNumber{Q}) -> CliffordNumber{Q}

Normalizes x so that its magnitude (as calculated by abs2(x)) is 1.

CliffordNumbers.odious_number — Method

CliffordNumbers.odious_number(n::Integer)

Returns the nth odious number, with the first odious number (n == 1) defined to be 1.

Odious numbers are numbers which have an odd Hamming weight (sum of its binary digits).

CliffordNumbers.right_contraction — Method

right_contraction(x::CliffordNumber{Q}, y::CliffordNumber{Q}) -> CliffordNumber{Q}

Calculates the right contraction of x and y.

For basis blades A of grade m and B of grade n, the right contraction is zero if m < n, otherwise it is grade_select(A*B, m-n).

CliffordNumbers.sandwich — Method

sandwich(x::CliffordNumber{Q}, y::CliffordNumber{Q})

Calculates the sandwich product of x with y: ~y * x * y, but with corrections for numerical stability.

CliffordNumbers.scalar_product — Method

scalar_product(x::CliffordNumber{Q,T1}, y::CliffordNumber{Q,T2}) -> promote_type(T1,T2)

Calculates the scalar product of two Clifford numbers with quadratic form Q. The result is a Real or Complex number. This can be converted back to a CliffordNumber.

This is equal to grade_select(x*y, 0) but is significantly more efficient.

CliffordNumbers.select_grade — Method

select_grade(x::CliffordNumber, g::Integer)

Returns a multivector similar to x where all elements not of grade g are equal to zero.

CliffordNumbers.sign_of_mult — Method

CliffordNumbers.sign_of_mult(a::T, b::T) where T<:BitIndex{QuadraticForm{P,Q,R}} -> Int8

Returns an Int8 that carries the sign associated with the multiplication of two basis blades of Clifford/geometric algebras of the same quadratic form.

CliffordNumbers.signbit_of_mult — Method

CliffordNumbers.signbit_of_mult(a::Integer, [b::Integer]) -> Bool
CliffordNumbers.signbit_of_mult(a::BitIndex, [b::BitIndex]) -> Bool

Calculates the sign bit associated with multiplying basis elements indexed with bit indices supplied as either integers or BitIndex instances. The sign bit flips when the order of a and b are reversed, unless a === b.

As with Base.signbit(), true represents a negative sign and false a positive sign. However, in degenerate metrics (such as those of projective geometric algebras) the sign bit may be irrelevant as the multiplication of those basis blades would result in zero.

CliffordNumbers.signmask — Function

CliffordNumbers.signmask(T::Type{<:Integer}, signbit::Bool = true) -> T1

Generates a signmask, or a string of bits where the only 1 bit is the sign bit. If signbit is set to false, this returns a string of bits.

CliffordNumbers.subscript_string — Method

CliffordNumbers.subscript_string(x::Number) -> String

Produces a string representation of a number in subscript format.

CliffordNumbers.undual — Method

undual(x::CliffordNumber) -> CliffordNumber

Calculates the undual of x, which is equal to the left contraction of x with the pseudoscalar. This function can be used to reverse the behavior of dual().

CliffordNumbers.versor_inverse — Method

versor_inverse(x::CliffordNumber)

Calculates the versor inverse of x, equal to x / scalar_product(x, ~x), so that x * inv(x) == inv(x) * x == 1.

The versor inverse is only guaranteed to be an inverse for blades and versors. Not all Clifford numbers have a well-defined inverse, since Clifford numbers have zero divisors (for instance, in the algebra of physical space, 1 + e₁ has a zero divisor).

CliffordNumbers.wedge — Method

wedge(x::CliffordNumber{Q}, y::CliffordNumber{Q}) -> CliffordNumber{Q}

Calculates the wedge (outer) product of two Clifford numbers x and y with quadratic form Q.