Extensions

[Quaternions.jl]

Quaternions.Quaternion — Method

Quaternion(c::AbstractCliffordNumber{VGA(3)})
Quaternion{T}(c::AbstractCliffordNumber{VGA(3)})

Constructs a quaternion from an element of the algebra of physical space, the 3D geometric algebra with a positive-definite signature whose even subalgebra is isomorphic to the quaternion algebra ℍ. Any odd-grade coefficients of c are lost. If the type parameter T is supplied, the scalars of the input are converted to type T.

If loss of odd-grade coefficients should throw an error, use convert(Quaternion, c) or convert(Quaternion{T}, c) instead of the constructor.

Quaternions.slerp — Method

slerp(x::AbstractCliffordNumber{VGA(3)}, y::AbstractCliffordNumber{VGA(3)}, t::Real)
    -> EvenCliffordNumber{VGA(3)}

Performs spherical linear interpolation for rotors in the 3D vanilla geometric algebra, treating them as if they were Quaternion instances.

Quaternions.slerp — Method

slerp(x::Quaternion, y::AbstractCliffordNumber{VGA(3)}, t::Real) -> EvenCliffordNumber{VGA(3)}
slerp(x::AbstractCliffordNumber{VGA(3)}, y::Quaternion, t::Real) -> EvenCliffordNumber{VGA(3)}

Performs spherical linear interpolation for rotors in the 3D vanilla geometric algebra, treating them as if they were Quaternion instances. The resulting output gains Clifford number semantics, so it is of type EvenCliffordNumber{VGA(3)} instead of Quaternion.

[Quaternions.jl]: https://github.com/JuliaGeometry/Quaternions.jl