Quadratic forms

QuadraticForm type and aliases

CliffordNumbers.QuadraticForm — Type

CliffordNumbers.QuadratricForm

Represents a quadratric with P dimensions which square to +1, Q dimensions which square to -1, and R dimensions which square to 0, in that order.

By convention, this type is used as a tag, and is never instantiated.

CliffordNumbers.QFReal — Type

CliffordNumbers.QFComplex

The quadratic form with zero dimensions, QuadraticForm{0,0,0}, isomorphic to the real numbers.

CliffordNumbers.QFComplex — Type

CliffordNumbers.QFComplex

The quadratic form with one dimension squaring to -1, QuadraticForm{0,1,0}. This generates a Clifford algebra isomorphic to the complex numbers.

CliffordNumbers.QFPositiveDefinite — Type

CliffordNumbers.QFPositiveDefinite{P} (alias for QuadraticForm{P,0,0})

A positive-definite quadratic form with P dimensions.

CliffordNumbers.QFNondegenerate — Type

CliffordNumbers.QFNondegenerate{P,Q} (alias for QuadraticForm{P,Q,0})

Represents a non-degenerate quadratic form (one without dimensions which square to 0).

CliffordNumbers.QFExterior — Type

CliffordNumbers.QFExterior{R} (alias for QuadraticForm{0,0,R})

The quadratic form associated with an exterior algebra, which can be thought of as a Clifford algebra where all dimensions are degenerate.

CliffordNumbers.APS — Type

APS

The algebra of physical space, Cl(3,0,0). An alias for QuadraticForm{3,0,0}.

CliffordNumbers.STA — Type

STA

Spacetime algebra with a mostly negative signature (particle physicist's convention), Cl(1,3,0). An alias for QuadraticForm{1,3,0}.

The negative signature is used by default to distinguish this algebra from conformal geometric algebras, which use a mostly positive signature by convention.

CliffordNumbers.VGA — Function

VGA(D) -> Type{QuadraticForm{D,0,0}}

Creates the type of a quadratic form associated with a vector/vanilla geometric algebra (VGA) of dimension D.

For reasons of type stability, avoid calling this function without constant arguments.

CliffordNumbers.PGA — Function

PGA(D) -> Type{QuadraticForm{D,0,1}}

Creates the type of a quadratic form associated with a projective geometric algebra (PGA) of dimension D.

For reasons of type stability, avoid calling this function without constant arguments.

CliffordNumbers.CGA — Function

CGA(D) -> Type{QuadraticForm{D+1,1,0}}

Creates the type of a quadratic form associated with a conformal geometric algebra (CGA) of dimension D.

For reasons of type stability, avoid calling this function without constant arguments.

Associated methods

CliffordNumbers.dimension — Function

dimension(::Type{QuadraticForm{P,Q,R}}) -> P + Q + R

Returns the number of dimensions associated with a Clifford algebra.

Note that this may not be the same as the number of dimensions being modeled by the algebra: for instance, projective geometric algebras have one more dimension than the space they are modeling.

CliffordNumbers.elements — Function

elements(Q::Type{QuadraticForm}) -> 2^dimension(Q)

Returns the number of basis elements associated with Q.

CliffordNumbers.grades — Function

grades(Q::Type{QuadraticForm}) -> 0:dimension(Q)

Returns the number of grades associated with Q.

Base.sign — Method

sign(::Type{QuadraticForm{P,Q,R}}, i::Integer) -> Int8

Gets the sign associated with dimension i of a quadratric form.