CliffordNumbers

Documentation for CliffordNumbers.

CliffordNumbers.APS

CliffordNumbers.BitIndex

CliffordNumbers.BitIndices

CliffordNumbers.CliffordNumber

CliffordNumbers.PGA2D

CliffordNumbers.PGA3D

CliffordNumbers.QuadraticForm

CliffordNumbers.STAminus

CliffordNumbers.STAplus

Base.:*

Base.:*

Base.sign

CliffordNumbers.dot

CliffordNumbers.elementwise_product

CliffordNumbers.hamming_weight

CliffordNumbers.isevil

CliffordNumbers.sign_of_mult

CliffordNumbers.subscript_string

CliffordNumbers.wedge

CliffordNumbers.APS — Type

APS

The algebra of physical space, Cl(3,0,0). An alias for QuadraticForm{3,0,0}.

CliffordNumbers.BitIndex — Type

CliffordNumbers.BitIndex{Cl<:QuadraticForm}

An index corresponding to a signed basis element of a Clifford algebra.

The basis elements of a Clifford number may be described with a binary word with length equal to the dimension of the space. The nth bit corresponds to whether the basis element e_n is used to construct the element with the given index.

As an example, in the algebra of physical space, the element e₁ can be represented with the binary number 0b001, e₂ with 0b010, and e₃ with 0b100. Elements of higher grades can be accessed by using the xor operation: for example, the basis bivector e₁e₃ can be represented with xor(0b001, 0b100) == 0b101. The scalar element is always represented with 0.

The sign of the index is stored. A negative sign corresponds to an odd permutation of the basis elements, and

The QuadraticForm is a tag that marks the Clifford algebra associated with the index. All operations on such BitIndex only depend on the value modulo elements(Cl). If a generic index is desired, QuadraticForm may be used as the tag.

CliffordNumbers.BitIndices — Type

BitIndices{Cl<:QuadraticForm}

Stores the range of valid BitIndices for the Clifford algebra with quadratic form Cl.

CliffordNumbers.CliffordNumber — Type

CliffordNumber{Cl,T,L}

An element of a Clifford algebra, often referred to as a multivector, with quadratic form Cl, element type T, and length L (which depends entirely on Cl).

The coefficients are ordered by taking advantage of the natural binary structure of the basis. The grade of an element is given by the Hamming weight of its index. For the algebra of physical space, the order is: 1, e₁, e₂, e₁₂, e₃, e₁₃, e₂₃, e₁₂₃ = i. This order allows for more aggressive SIMD optimization when calculating the geometric product.

CliffordNumbers.PGA2D — Type

PGA3D

Projective geometric algebra of 2 dimensions and 1 projective dimension, Cl(2,0,1). An alias for QuadraticForm{2,0,1}.

CliffordNumbers.PGA3D — Type

PGA3D

Projective geometric algebra of 3 dimensions and 1 projective dimension, Cl(3,0,1). An alias for QuadraticForm{3,0,1}.

CliffordNumbers.QuadraticForm — Type

CliffordNumbers.QuadratricForm

Represents a quadratric with P dimensions which square to +1, Q dimensions which square to -1, and R dimensions which square to 0, in that order.

By convention, this type is used as a tag, and is never instantiated.

CliffordNumbers.STAminus — Type

STAminus

Spacetime algebra with a mostly negative signature (particle physicist's convention), Cl(1,3,0). An alias for QuadraticForm{1,3,0}.

CliffordNumbers.STAplus — Type

STAplus

Spacetime algebra with a mostly positive signature (mathematician's convention), Cl(3,1,0). An alias for QuadraticForm{3,1,0}.

Base.:* — Method

*(i1::BitIndex{Cl}, i2::BitIndex{Cl}) -> BitIndex{Cl}

Returns the basis element that results from the multiplication of basis elements i1 and i2.

Base.:* — Method

*(m1::CliffordNumber{Cl}, m2::CliffordNumber{Cl}) -> CliffordNumber{Cl}

Calculates the geometric product between multivectors/Clifford numbers m1 and m2 which share the quadratic form Cl.

Base.sign — Method

sign(::Type{QuadraticForm{P,Q,R}}, i::Integer) -> Int8

Gets the sign associated with dimension i of a quadratric form.

CliffordNumbers.dot — Method

dot(m1::CliffordNumber{Cl}, m2::CliffordNumber{Cl}) -> Number
m1 · m2 -> CliffordNumber{Cl}

Calculates the dot (inner) product of two Clifford numbers with quadratic form Cl. The result is a Real or Complex number. This can be converted back to a CliffordNumber.

CliffordNumbers.elementwise_product — Method

CliffordAlgebra.elementwise_product(
    m1::CliffordNumber{Cl},
    m2::CliffordNumber{Cl},
    i1::Integer,
    i2::Integer
)

Calculates the geometric product between element i1 of Clifford number m1 and element i2 of Clifford number m2.

CliffordNumbers.hamming_weight — Method

CliffordNumbers.hamming_weight(i::Integer) -> Int

Calculates the Hamming weight of an integer. This is used to determine the grade of a component of a StaticMultivector.

CliffordNumbers.isevil — Method

CliffordNumbers.isevil(i::Integer) -> Bool

Determines whether a number is evil, meaning that its Hamming weight (sum of its binary digits) is even.

CliffordNumbers.sign_of_mult — Method

CliffordNumbers.sign_of_mult([Cl::Type{QuadraticForm{P,Q,R}}], a::Integer, b::Integer)

Calculates the sign associated with multiplying basis elements indexed with bit indices supplied as integers. The sign reverses whenever the order of a and b are reverse, provided they are not the same.

CliffordNumbers.subscript_string — Method

CliffordNumbers.subscript_string(x::Number) -> String

Produces a string representation of a number in subscript format.

CliffordNumbers.wedge — Method

wedge(m1::CliffordNumber{Cl}, m2::CliffordNumber{Cl}) -> CliffordNumber{Cl}

Calculates the wedge (outer) product of two Clifford numbers with quadratic form Cl. The result is another CliffordNumber{Cl}.