CliffordNumbers

CliffordNumbers.jl is a package that provides fully static multivectors (Clifford numbers) in arbitrary dimensions and metrics. While in many cases, sparse representations of multivectors are more efficient, for spaces of low dimension, dense static representations may provide a performance and convenience advantage.

CliffordNumbers.APS

CliffordNumbers.AbstractBitIndices

CliffordNumbers.AbstractCliffordNumber

CliffordNumbers.BaseNumber

CliffordNumbers.BitIndex

CliffordNumbers.BitIndex

CliffordNumbers.BitIndices

CliffordNumbers.CliffordNumber

CliffordNumbers.KVector

CliffordNumbers.QFComplex

CliffordNumbers.QFReal

CliffordNumbers.QuadraticForm

CliffordNumbers.RepresentedGrades

CliffordNumbers.STA

CliffordNumbers.TransformedBitIndices

CliffordNumbers.Z2CliffordNumber

Base.:*

Base.abs

Base.abs2

Base.conj

Base.exp

Base.real

Base.reverse

Base.sign

CliffordNumbers.CGA

CliffordNumbers.PGA

CliffordNumbers.VGA

CliffordNumbers._sort_with_parity!

CliffordNumbers.check_element_count

CliffordNumbers.commutator

CliffordNumbers.dot

CliffordNumbers.dual

CliffordNumbers.elementwise_product

CliffordNumbers.evil_number

CliffordNumbers.exppi

CliffordNumbers.exptau

CliffordNumbers.grade_involution

CliffordNumbers.hamming_number

CliffordNumbers.has_grades_of

CliffordNumbers.has_wedge

CliffordNumbers.hestenes_product

CliffordNumbers.isevil

CliffordNumbers.isodious

CliffordNumbers.ispseudoscalar

CliffordNumbers.isscalar

CliffordNumbers.next_of_hamming_weight

CliffordNumbers.nonzero_grades

CliffordNumbers.normalize

CliffordNumbers.number_of_parity

CliffordNumbers.numeric_type

CliffordNumbers.odious_number

CliffordNumbers.promote_numeric_type

CliffordNumbers.sandwich

CliffordNumbers.scalar_product

CliffordNumbers.select_grade

CliffordNumbers.sign_of_mult

CliffordNumbers.signbit_of_mult

CliffordNumbers.signmask

CliffordNumbers.similar_type

CliffordNumbers.subscript_string

CliffordNumbers.to_basis_str

CliffordNumbers.undual

CliffordNumbers.versor_inverse

CliffordNumbers.wedge

CliffordNumbers.BaseNumber — Type

CliffordNumbers.BaseNumber

Union of subtypes of Number provided in the Julia Base module: Real and Complex. This encompasses all types that may be used to construct a CliffordNumber.

CliffordNumbers.APS — Type

APS

The algebra of physical space, Cl(3,0,0). An alias for QuadraticForm{3,0,0}.

CliffordNumbers.AbstractBitIndices — Type

AbstractBitIndices{Q,C<:AbstractCliffordNumber{Q}} <: AbstractVector{BitIndex{Q}}

Supertype for vectors containing all valid BitIndex{Q} objects for the basis elements represented by C.

CliffordNumbers.AbstractCliffordNumber — Type

AbstractCliffordNumber{Q,T} <: Number

An element of a Clifford algebra, often referred to as a multivector, with quadratic form Q and element type T.

Interface

Required implementation

All subtypes C of AbstractCliffordNumber{Q} must implement the following functions:

Base.length(x::C) should return the number of nonzero basis elements represented by x.

CliffordNumbers.similar_type(::Type{C}, ::Type{T}, ::Type{Q}) where {C,T,Q} should construct a

new type similar to C which subtypes AbstractCliffordNumber{Q,T} that may serve as a constructor.

Base.getindex(x::C, b::BitIndex{Q}) should allow one to recover the coefficients associated

with each basis blade represented by C.

Required implementation for static types

Base.length(::Type{C}) should be defined, with Base.length(x::C) = length(typeof(x)).

Base.Tuple(x::C) should return the tuple used to construct x. The fallback is

getfield(x, :data)::Tuple, so any type declared with a NTuple field named data should have this defined automatically.

CliffordNumbers.BitIndex — Type

BitIndex{Q<:QuadraticForm}

A representation of an index corresponding to a basis blade of the geometric algebra with quadratic form Q.

CliffordNumbers.BitIndex — Method

BitIndex{Q}(x::Integer...)
BitIndex{Q}(v::AbstractVector{<:Integer})

Constructs a BitIndex{Q} from a list of integers that represent the basis vectors of the space.

This package uses a lexicographic convention for basis blades: in the algebra of physical space, the basis bivectors are {e₁e₂, e₁e₃, e₂e₃}. The sign of the BitIndex{Q} is negative when the parity of the basis vector permutation is odd.

CliffordNumbers.BitIndices — Type

BitIndices{Q<:QuadraticForm,C<:AbstractCliffordNumber{Q,<:Any}} <: AbstractVector{BitIndex{Q}}

Represents a range of valid BitIndex objects for the nonzero components of a given multivector with quadratic form Q.

For a generic AbstractCliffordNumber{Q}, this returns BitIndices{CliffordNumber{Q}}, which contains all possible indices for a multivector associated with the quadratic form Q. This may also be constructed with BitIndices(Q).

For sparse representations, such as KVector{K,Q}, the object only contains the indices of the nonzero elements of the multivector.

Construction

BitIndices can be constructed by calling the type constructor with either the multivector or its type.

Indexing

BitIndices always uses one-based indexing like most Julia arrays. Although it is more natural in the dense case to use zero-based indexing, as the basis blades are naturally encoded in the indices for the dense representation of CliffordNumber, one-based indexing is used by the tuples which contain the data associated with this package's implementations of Clifford numbers.

Interfaces for new subtypes of AbstractCliffordNumber

When defining the behavior of BitIndices for new subtypes T of AbstractCliffordNumber, Base.getindex(::BitIndices{Q,T}, i::Integer) should be defined so that all indices of T that are not constrained to be zero are returned.

CliffordNumbers.CliffordNumber — Type

CliffordNumber{Q,T,L} <: AbstractCliffordNumber{Q,T}

A dense multivector (or Clifford number), with quadratic form Q, element type T, and length L (which depends entirely on Q).

The coefficients are ordered by taking advantage of the natural binary structure of the basis. The grade of an element is given by the Hamming weight of its index. For the algebra of physical space, the order is: 1, e₁, e₂, e₁₂, e₃, e₁₃, e₂₃, e₁₂₃ = i. This order allows for more aggressive SIMD optimization when calculating the geometric product.

CliffordNumbers.KVector — Type

KVector{K,Q,T,L} <: AbstractCliffordNumber{Q,T}

A multivector consisting only linear combinations of basis blades of grade K - in other words, a k-vector.

k-vectors have binomial(dimension(Q), K) components.

CliffordNumbers.QFComplex — Type

CliffordNumbers.QFComplex

The quadratic form with one dimension squaring to -1, QuadraticForm{0,1,0}. This generates a Clifford algebra isomorphic to the complex numbers.

CliffordNumbers.QFReal — Type

CliffordNumbers.QFComplex

The quadratic form with zero dimensions, QuadraticForm{0,0,0}, isomorphic to the real numbers.

CliffordNumbers.QuadraticForm — Type

CliffordNumbers.QuadratricForm

Represents a quadratric with P dimensions which square to +1, Q dimensions which square to -1, and R dimensions which square to 0, in that order.

By convention, this type is used as a tag, and is never instantiated.

CliffordNumbers.RepresentedGrades — Type

RepresentedGrades{C<:AbstractCliffordNumber} <: AbstractVector{Bool}

A vector describing the grades explicitly represented by some Clifford number type C.

Indexing of a RepresentedGrades{Q} is zero-based, with indices corresponding to all grades from 0 (scalar) to dimension(Q) (pseudoscalar), with true values indicating nonzero grades, and false indicating that the grade is zero.

Implementation

The function nonzero_grades(::Type{T}) should be implemented for types T descending from AbstractCliffordNumber.

CliffordNumbers.STA — Type

STA

Spacetime algebra with a mostly negative signature (particle physicist's convention), Cl(1,3,0). An alias for QuadraticForm{1,3,0}.

The negative signature is used by default to distinguish this algebra from conformal geometric algebras, which use a mostly positive signature by convention.

CliffordNumbers.TransformedBitIndices — Type

TransformedBitIndices{Q,C,F} <: AbstractBitIndices{Q,C}

Lazy representation of BitIndices{Q,C} with some function of type f applied to each element. These objects can be used to perform common operations which act on basis blades or grades, such as the reverse or grade involution.

CliffordNumbers.Z2CliffordNumber — Type

CliffordNumbers.Z2CliffordNumber{P,Q,T,L}

A Clifford number whose only nonzero grades are even or odd. Clifford numbers of this form naturally arise as versors, the geometric product of 1-vectors.

The type parameter P is constrained to be a Bool: true for odd grade Clifford numbers, and false for even grade Clifford numbers, corresponding to the Boolean result of each grade modulo 2.

Type aliases

This type is not exported, and usually you will want to refer to the following aliases:

const EvenCliffordNumber{Q,T,L} = Z2CliffordNumber{false,Q,T,L}
const OddCliffordNumber{Q,T,L} = Z2CliffordNumber{true,Q,T,L}

Base.:* — Method

*(x::AbstractCliffordNumber{Q}, y::AbstractCliffordNumber{Q})
(x::AbstractCliffordNumber{Q})(y::AbstractCliffordNumber{Q})

Calculates the geometric product of x and y, returning the smallest type which is able to represent all nonzero basis blades of the result.

Base.abs — Method

abs2(x::CliffordNumber{Q,T}) -> Union{Real,Complex}

Calculates the norm of x, equal to sqrt(scalar_product(x, ~x)).

Base.abs2 — Method

abs2(x::AbstractCliffordNumber{Q,T}) -> T

Calculates the squared norm of x, equal to scalar_product(x, ~x).

Base.conj — Method

conj(b::BitIndex) -> BitIndex
conj(x::AbstractCliffordNumber{Q,T}) -> typeof(x)

Calculates the Clifford conjugate of the basis blade indexed by b or the Clifford number x. This is equal to grade_involution(reverse(x)).

Base.exp — Method

exp(x::CliffordNumber{Q}) -> CliffordNumber{Q,<:AbstractFloat}

Returns the natural exponential of a Clifford number.

For special cases where m squares to a scalar, the following shortcuts can be used to calculate exp(x):

When x^2 < 0: exp(x) === cos(abs(x)) + x * sin(abs(x)) / abs(x)

When x^2 > 0: exp(x) === cosh(abs(x)) + x * sinh(abs(x)) / abs(x)

When x^2 === 0: exp(x) == 1 + x

See also: exppi, exptau.

Base.real — Method

real(x::CliffordNumber{Q,T<:Real}) = T

Return the real (scalar) portion of a real Clifford number.

Base.reverse — Method

reverse(b::BitIndex) -> BitIndex
reverse(x::AbstractCliffordNumber{Q,T}) -> typeof(x)

Performs the reverse operation on the basis blade indexed by b or the Clifford number x. The sign of the reverse depends on the grade, and is positive for g % 4 in 0:1 and negative for g % 4 in 2:3.

Base.sign — Method

sign(::Type{QuadraticForm{P,Q,R}}, i::Integer) -> Int8

Gets the sign associated with dimension i of a quadratric form.

CliffordNumbers.CGA — Method

CGA(D) -> Type{QuadraticForm{D+1,1,0}}

Creates the type of a quadratic form associated with a conformal geometric algebra (CGA) of dimension D.

For reasons of type stability, avoid calling this function without constant arguments.

CliffordNumbers.PGA — Method

PGA(D) -> Type{QuadraticForm{D,0,1}}

Creates the type of a quadratic form associated with a projective geometric algebra (PGA) of dimension D.

For reasons of type stability, avoid calling this function without constant arguments.

CliffordNumbers.VGA — Method

VGA(D) -> Type{QuadraticForm{D,0,0}}

Creates the type of a quadratic form associated with a vector/vanilla geometric algebra (VGA) of dimension D.

For reasons of type stability, avoid calling this function without constant arguments.

CliffordNumbers._sort_with_parity! — Method

CliffordNumbers._sort_with_parity!(v::AbstractVector{<:Real}) -> Tuple{typeof(v),Bool}

Performs a parity-tracking insertion sort of v, which modifies v in place. The function returns a tuple containing v and the parity, which is true for an odd permutation and false for an even permutation. This is implemented with a modified insertion sort algorithm.

CliffordNumbers.check_element_count — Method

CliffordNumbers.check_element_count(f, Q::Type{<:QuadraticForm}, [L,] data)

Ensures that the number of elements in data is the same as the result of f(Q), where f is a function that generates the expected number of elements for the type. This function is used in the inner constructors of subtypes of AbstractCliffordNumber{Q} to ensure that the input has the correct length.

If provided, the length type parameter L can be included as an argument, and it will be checked for type (must be an Int) and value (must be equal to f(Q)).

This function returns nothing, but throws an AssertionError for failed checks.

CliffordNumbers.commutator — Method

commutator(x::AbstractCliffordNumber{Q}, y::AbstractCliffordNumber{Q})
×(x::AbstractCliffordNumber{Q}, y::AbstractCliffordNumber{Q})

Calculates the commutator product, equal to 1//2 * (x*y - y*x), or equivalently, 1//2 * (x*y - reverse(x*y)).

CliffordNumbers.dot — Method

dot(x::CliffordNumber{Q}, y::CliffordNumber{Q}) -> CliffordNumber{Q}

Calculates the dot product of x and y.

For basis blades A of grade m and B of grade n, the dot product is equal to the left contraction when m >= n and is equal to the right contraction when n >= m.

CliffordNumbers.dual — Method

dual(x::CliffordNumber) -> CliffordNumber

Calculates the dual of x, which is equal to the left contraction of x with the inverse of the pseudoscalar. However,

Note that the dual has some properties that depend on the dimension and quadratic form:

The inverse of the unit pseudoscalar depends on the dimension of the space. Therefore, the

periodicity of

If the metric is degenerate, the dual is not unique.

CliffordNumbers.elementwise_product — Method

CliffordNumbers.elementwise_product(
    [::Type{C},]
    x::AbstractCliffordNumber{Q},
    y::AbstractCliffordNumber{Q},
    a::BitIndex{Q},
    b::BitIndex{Q},
    [condition = true]
) where {Q,T<:AbstractCliffordNumber{Q}} -> C

Calculates the geometric product between the element of x indexed by a and the element of y indexed by b. The result is returned as type C, but this can be inferred automatically if not provided.

An optional boolean condition can be provided, which simplifies the implementation of certain products derived from the geometric product.

CliffordNumbers.evil_number — Method

CliffordNumbers.evil_number(n::Integer)

Returns the nth evil number, with the first evil number (n == 1) defined to be 0.

Evil numbers are numbers which have an even Hamming weight (sum of its binary digits).

CliffordNumbers.exppi — Method

exppi(x::CliffordNumber)

Returns the natural exponential of π * x with greater accuracy than exp(π * x) in the case where x^2 is a negative scalar.

See also: exp, exptau.

CliffordNumbers.exptau — Method

exptau(x::CliffordNumber)

Returns the natural exponential of 2π * x with greater accuracy than exp(2π * x) in the case where x^2 is a negative scalar.

See also: exp, exppi.

CliffordNumbers.grade_involution — Method

grade_involution(b::BitIndex) -> BitIndex
grade_involution(x::AbstractCliffordNumber{Q,T}) -> typeof(x)

Calculates the grade involution of the basis blade indexed by b or the Clifford number x. This effectively reflects all of the basis vectors of the space along their own mirror operation, which makes elements of odd grade flip sign.

CliffordNumbers.hamming_number — Method

CliffordNumbers.hamming_number(w::Integer, n::Integer)

Gets the nth number with Hamming weight w. The first number with this Hamming weight (n = 1) is 2^w - 1.

Example

julia> CliffordNumbers.hamming_number(3, 2)
11

CliffordNumbers.has_grades_of — Method

has_grades_of(S::Type{<:AbstractCliffordNumber}, T::Type{<:AbstractCliffordNumber}) -> Bool
has_grades_of(x::AbstractCliffordNumber, y::AbstractCliffordNumber) -> Bool

Returns true if the grades represented in S are also represented in T; false otherwise.

CliffordNumbers.has_wedge — Method

CliffordNumbers.has_wedge(a::BitIndex{Q}, b::BitIndex{Q}) -> Bool

Returns true if the basis blades indexed by a and b may have a nonzero wedge product.

CliffordNumbers.hestenes_product — Method

hestenes_product(x::CliffordNumber{Q}, y::CliffordNumber{Q}) -> CliffordNumber{Q}

Returns the Hestenes product: this is equal to the dot product given by dot(x, y) but is equal to to zero when either x or y is a scalar.

CliffordNumbers.isevil — Method

CliffordNumbers.isevil(i::Integer) -> Bool

Determines whether a number is evil, meaning that its Hamming weight (sum of its binary digits) is even.

CliffordNumbers.isodious — Method

CliffordNumbers.isodious(i::Integer) -> Bool

Determines whether a number is odious, meaning that its Hamming weight (sum of its binary digits) is odd.

CliffordNumbers.ispseudoscalar — Method

ispseudoscalar(m::CliffordNumber)

Determines whether the Clifford number m is a pseudoscalar, meaning that it has no components with grades below the one equal to the dimension of the space.

CliffordNumbers.isscalar — Method

isscalar(m::CliffordNumber)

Determines whether the Clifford number m is a scalar, meaning that it has no components with grades above zero.

CliffordNumbers.next_of_hamming_weight — Method

CliffordNumbers.next_of_hamming_weight(n::Integer)

Returns the next integer with the same Hamming weight as n.

CliffordNumbers.nonzero_grades — Method

nonzero_grades(::Type{<:AbstractCliffordNumber})
nonzero_grades(::AbstractCliffordNumber)

A function returning an indexable object representing all nonzero grades of a Clifford number representation.

This function is used to define the indexing of RepresentedGrades, and should be defined for any subtypes of AbstractCliffordNumber.

Examples

julia> CliffordNumbers.nonzero_grades(CliffordNumber{APS})
0:3

julia> CliffordNumbers.nonzero_grades(KVector{2,APS})
2:2

CliffordNumbers.normalize — Method

normalize(x::AbstractCliffordNumber{Q}) -> AbstractCliffordNumber{Q}

Normalizes x so that its magnitude (as calculated by abs2(x)) is 1.

CliffordNumbers.number_of_parity — Method

CliffordNumbers.number_of_parity(n::Integer, modulo::Bool)

Returns the nth number whose Hamming weight is even (for modulo = false) or odd (for modulo = true).

CliffordNumbers.numeric_type — Method

numeric_type(::Type{<:AbstractCliffordNumber{Q,T}}) = T
numeric_type(T::Type{<:Union{Real,Complex}}) = T
numeric_type(x) = numeric_type(typeof(x))

Returns the numeric type associated with an AbstractCliffordNumber instance. For subtypes of Real and Complex, or their instances, this simply returns the input type or instance type.

Why not define eltype?

AbstractCliffordNumber instances behave like numbers, not arrays. If collect() is called on a Clifford number of type T, it should not construct a vector of coefficients; instead it should return an Array{T,0}. Similarly, a broadcasted multiplication should return the same result as normal multiplication, as is the case with complex numbers.

For subtypes T of Number, eltype(T) === T, and this is true for AbstractCliffordNumber.

CliffordNumbers.odious_number — Method

CliffordNumbers.odious_number(n::Integer)

Returns the nth odious number, with the first odious number (n == 1) defined to be 1.

Odious numbers are numbers which have an odd Hamming weight (sum of its binary digits).

CliffordNumbers.promote_numeric_type — Method

promote_numeric_type(x, y)

Calls promote_type() for the result of numeric_type() called on all arguments. For incompletely specified types, the result of numeric_type() is replaced with Bool, which always promotes to any larger numeric type.

CliffordNumbers.sandwich — Method

sandwich(x::CliffordNumber{Q}, y::CliffordNumber{Q})

Calculates the sandwich product of x with y: ~y * x * y, but with corrections for numerical stability.

CliffordNumbers.scalar_product — Method

scalar_product(x::CliffordNumber{Q,T1}, y::CliffordNumber{Q,T2}) -> promote_type(T1,T2)

Calculates the scalar product of two Clifford numbers with quadratic form Q. The result is a Real or Complex number. This can be converted back to an AbstractCliffordNumber.

CliffordNumbers.select_grade — Method

select_grade(x::CliffordNumber, g::Integer)

Returns a multivector similar to x where all elements not of grade g are equal to zero.

CliffordNumbers.sign_of_mult — Method

CliffordNumbers.sign_of_mult(a::T, b::T) where T<:BitIndex{QuadraticForm{P,Q,R}} -> Int8

Returns an Int8 that carries the sign associated with the multiplication of two basis blades of Clifford/geometric algebras of the same quadratic form.

CliffordNumbers.signbit_of_mult — Method

CliffordNumbers.signbit_of_mult(a::Integer, [b::Integer]) -> Bool
CliffordNumbers.signbit_of_mult(a::BitIndex, [b::BitIndex]) -> Bool

Calculates the sign bit associated with multiplying basis elements indexed with bit indices supplied as either integers or BitIndex instances. The sign bit flips when the order of a and b are reversed, unless a === b.

As with Base.signbit(), true represents a negative sign and false a positive sign. However, in degenerate metrics (such as those of projective geometric algebras) the sign bit may be irrelevant as the multiplication of those basis blades would result in zero.

CliffordNumbers.signmask — Function

CliffordNumbers.signmask(T::Type{<:Integer}, signbit::Bool = true) -> T1

Generates a signmask, or a string of bits where the only 1 bit is the sign bit. If signbit is set to false, this returns a string of bits.

CliffordNumbers.similar_type — Method

CliffordNumbers.similar_type(
    C::Type{<:AbstractCliffordNumber},
    [N::Type{<:BaseNumber} = numeric_type(C)],
    [Q::Type{<:QuadraticForm} = QuadraticForm(C)]
) -> Type{<:AbstractCliffordNumber{Q,N}}

Constructs a type similar to T but with numeric type N and quadratic form Q.

This function must be defined with all its arguments for each concrete type subtyping AbstractCliffordNumber.

CliffordNumbers.subscript_string — Method

CliffordNumbers.subscript_string(x::Number) -> String

Produces a string representation of a number in subscript format.

CliffordNumbers.to_basis_str — Method

CliffordNumbers.to_basis_str(b::BitIndex; [label], [pseudoscalar])

Creates a string representation of the basis element given by b.

The label parameter determines the symbol used for every basis element. In general, this defaults to e, but a few special cases use different symbols:

For the algebra of physical space, σ is used.

For the spacetime algebras of either signature, γ is used.

If pseudoscalar is set, the pseudoscalar may be printed using a different symbol from the rest of the basis elements.

CliffordNumbers.undual — Method

undual(x::CliffordNumber) -> CliffordNumber

Calculates the undual of x, which is equal to the left contraction of x with the pseudoscalar. This function can be used to reverse the behavior of dual().

CliffordNumbers.versor_inverse — Method

versor_inverse(x::CliffordNumber)

Calculates the versor inverse of x, equal to x / scalar_product(x, ~x), so that x * inv(x) == inv(x) * x == 1.

The versor inverse is only guaranteed to be an inverse for blades and versors. Not all Clifford numbers have a well-defined inverse, since Clifford numbers have zero divisors (for instance, in the algebra of physical space, 1 + e₁ has a zero divisor).

CliffordNumbers.wedge — Method

wedge(x::CliffordNumber{Q}, y::CliffordNumber{Q}) -> CliffordNumber{Q}

Calculates the wedge (outer) product of two Clifford numbers x and y with quadratic form Q.